+++ test/utils/utils.test.js
@@ -7,7 +7,6 @@ import {
   determineOrderDir,
   fixLeafletIcons,
   rangeOf,
-  roundTen,
 } from '../../src/utils/utils';
 
 describe('utils', () => {
@@ -84,23 +83,6 @@ describe('utils', () => {
         'result_4',
         'result_5',
       ]);
-    });
-  });
-
-  describe('roundTen', () => {
-    it('rounds provided number to the next multiple of ten', () => {
-      const expectationsPairs = [
-        [ 10, 10 ],
-        [ 12, 20 ],
-        [ 158, 160 ],
-        [ 5, 10 ],
-        [ -42, -40 ],
-      ];
-
-      expect.assertions(expectationsPairs.length);
-      expectationsPairs.forEach(([ number, expected ]) => {
-        expect(roundTen(number)).toEqual(expected);
-      });
     });
   });
 });
